% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{colScale}
\alias{colScale}
\title{Normalize the columns of a given matrix}
\usage{
colScale(X)
}
\arguments{
\item{mat}{Matrix to normalize}
}
\value{
returns column-normalized matrix
}
\description{
Normalize the columns of a given matrix
}
