% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.sparseMatrix}
\alias{is.sparseMatrix}
\title{Matrix Crossproduct
Given a list of matrices x1, x2, ... , xn as arguments, return a matrix cross-product. 
x1%*%x2%*%xn}
\usage{
is.sparseMatrix(x)
}
\arguments{
\item{lst}{list of matrix}
}
\value{
returns crossproduct
}
\description{
Matrix Crossproduct
Given a list of matrices x1, x2, ... , xn as arguments, return a matrix cross-product. 
x1%*%x2%*%xn
}
