% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{L2Norm}
\alias{L2Norm}
\title{L2 normalize the columns (or rows) of a given matrix}
\usage{
L2Norm(mat, MARGIN = 1)
}
\arguments{
\item{mat}{Matrix to cosine normalize}

\item{MARGIN}{Perform normalization over rows (1) or columns (2)}
}
\value{
returns l2-normalized matrix
}
\description{
L2 normalize the columns (or rows) of a given matrix
}
