% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rowScale}
\alias{rowScale}
\title{Normalize the rows of a given matrix}
\usage{
rowScale(X)
}
\arguments{
\item{mat}{Matrix to normalize}
}
\value{
returns row-normalized matrix
}
\description{
Normalize the rows of a given matrix
}
